Move Field(s) to common. (#5271)
Make type Field(s) reusable and seperate from transformation logic.
